Update `UserTask` controller route

+++ app/src/Controller/Api/UserTaskController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller\Api;
 
+use App\Entity\User;
 use App\Entity\UserTask;
 use App\Pagination\PaginationQuery;
 use App\Pagination\PaginationResult;
@@ -16,7 +17,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Validator\ConstraintViolationListInterface;
 
 /**
- * @Rest\Route("/user-task")
+ * @Rest\Route("/user/{user_id}/user-task")
  */
 class UserTaskController
 {
@@ -24,26 +25,32 @@ class UserTaskController
      * @Rest\Get("/list")
      * @Rest\View()
      *
+     * @Sensio\ParamConverter("user", options={"id" = "user_id"})
+     *
+     * @param User $user
      * @param PaginationQuery $paginationQuery
      * @param UserTaskService $userTaskService
      * @return PaginationResult
      */
-    public function list(PaginationQuery $paginationQuery, UserTaskService $userTaskService): PaginationResult
+    public function list(User $user, PaginationQuery $paginationQuery, UserTaskService $userTaskService): PaginationResult
     {
-        return $userTaskService->list($paginationQuery);
+        return $userTaskService->list($user, $paginationQuery);
     }
 
     /**
      * @Rest\Post("")
      *
+     * @Sensio\ParamConverter("user", options={"id" = "user_id"})
      * @Sensio\ParamConverter("request", converter="fos_rest.request_body")
      *
+     * @param User $user
      * @param CreateUserTaskRequest $request
      * @param ConstraintViolationListInterface $constraintViolationList
      * @param UserTaskService $userTaskService
      * @return View
      */
     public function create(
+        User $user,
         CreateUserTaskRequest $request,
         ConstraintViolationListInterface $constraintViolationList,
         UserTaskService $userTaskService
@@ -52,26 +59,37 @@ class UserTaskController
             return View::create($constraintViolationList, Response::HTTP_BAD_REQUEST);
         }
 
-        return View::create($userTaskService->create($request), Response::HTTP_CREATED);
+        return View::create($userTaskService->create($user, $request), Response::HTTP_CREATED);
     }
 
     /**
      * @Rest\Get("/{id}")
-     * @Rest\View()
      *
-     * @param UserTask $user
-     * @return UserTask
+     * @Sensio\ParamConverter("user", options={"id" = "user_id"})
+     *
+     * @param User $user
+     * @param UserTask $userTask
+     * @return View
      */
-    public function read(UserTask $user): UserTask
+    public function read(User $user, UserTask $userTask): View
     {
-        return $user;
+        if (!$userTask->getUser()->getId()->equals($user->getId())) {
+            return View::create(
+                ['error' => 'The given user id doesn\'t corresponds with user task user.'],
+                Response::HTTP_BAD_REQUEST
+            );
+        }
+
+        return View::create($userTask, Response::HTTP_OK);
     }
 
     /**
      * @Rest\Patch("/{id}")
      *
+     * @Sensio\ParamConverter("user", options={"id" = "user_id"})
      * @Sensio\ParamConverter("request", converter="fos_rest.request_body")
      *
+     * @param User $user
      * @param UserTask $userTask
      * @param UpdateUserTaskRequest $request
      * @param ConstraintViolationListInterface $constraintViolationList
@@ -79,6 +97,7 @@ class UserTaskController
      * @return View
      */
     public function update(
+        User $user,
         UserTask $userTask,
         UpdateUserTaskRequest $request,
         ConstraintViolationListInterface $constraintViolationList,
@@ -86,6 +105,13 @@ class UserTaskController
     ): View {
         if ($constraintViolationList->count() > 0) {
             return View::create($constraintViolationList, Response::HTTP_BAD_REQUEST);
+        }
+
+        if (!$userTask->getUser()->getId()->equals($user->getId())) {
+            return View::create(
+                ['error' => 'The given user id doesn\'t corresponds with user task user.'],
+                Response::HTTP_BAD_REQUEST
+            );
         }
 
         $userTaskService->update($userTask, $request);
@@ -96,12 +122,22 @@ class UserTaskController
     /**
      * @Rest\Delete("/{id}")
      *
+     * @Sensio\ParamConverter("user", options={"id" = "user_id"})
+     *
+     * @param User $user
      * @param UserTask $userTask
      * @param UserTaskService $userTaskService
      * @return View
      */
-    public function delete(UserTask $userTask, UserTaskService $userTaskService): View
+    public function delete(User $user, UserTask $userTask, UserTaskService $userTaskService): View
     {
+        if (!$userTask->getUser()->getId()->equals($user->getId())) {
+            return View::create(
+                ['error' => 'The given user id doesn\'t corresponds with user task user.'],
+                Response::HTTP_BAD_REQUEST
+            );
+        }
+
         $userTaskService->delete($userTask);
 
         return View::create(null, Response::HTTP_NO_CONTENT);
